feat: sync config store with backend API
