+++ minmmo/src/app/admin.main.tsx
@@ -5,7 +5,12 @@ import { AdminPortal } from '@cms/AdminPortal'
 import { load, subscribe } from '@config/store'
 import { rebuildFromConfig } from '@content/registry'
 
-rebuildFromConfig(load())
 subscribe(rebuildFromConfig)
+
+load()
+  .then(rebuildFromConfig)
+  .catch((error) => {
+    console.error('Failed to load configuration', error)
+  })
 
 ReactDOM.createRoot(document.getElementById('root')!).render(<AdminPortal />)
